+++ demo/src/objects/Simple.ts
@@ -1,0 +1,3 @@
+import { CanvasObject } from 'react-canvas-animate'
+
+export class Simple extends CanvasObject {}