+++ src/lib/CanvasObjectManager.ts
@@ -1,17 +1,28 @@
 import { CanvasContext, Context2D } from '../components/Canvas'
-import CanvasObject from './CanvasObject'
+import CanvasObject, {
+  CanvasObjectState,
+  CanvasObjectProps,
+} from './CanvasObject'
 
 class CanvasObjectManager<
   T extends CanvasObject = CanvasObject,
+  S extends CanvasObjectState = CanvasObjectState,
+  P extends CanvasObjectProps = CanvasObjectProps,
   C extends CanvasContext = Context2D,
 > {
   private objects: T[] = []
   private context: C
 
-  constructor(context: C, initialObjects: T[] = []) {
+  constructor(
+    context: C,
+    objects: [new (context: C, state: S) => T, S][] = [],
+  ) {
     this.context = context
-    this.objects = initialObjects
+    this.objects = this.createList(objects)
   }
+
+  // Stack methods
+  // ---------------
 
   add(object: T): void {
     this.objects.push(object)
@@ -40,21 +51,34 @@ class CanvasObjectManager<
     this.objects = []
   }
 
-  create(ObjectClass: new (context: C) => T): T {
-    const newObject = new ObjectClass(this.context)
+  // Factory method
+  // ---------------
+
+  createList(objects: [new (context: C, state: S) => T, S][]): T[] {
+    const newObjects = objects.map(([ObjectClass, state]) => {
+      const newObject = new ObjectClass(this.context, state)
+      this.add(newObject)
+      return newObject
+    })
+    return newObjects
+  }
+
+  create(ObjectClass: new (context: C, state: S) => T, state: S): T {
+    const newObject = new ObjectClass(this.context, state)
     this.add(newObject)
     return newObject
   }
 
   // Callbacks
+  // ---------------
 
-  update(..._args: any[]): void {
+  update(_args: P): void {
     for (const object of this.objects) {
       object.callUpdate(_args)
     }
   }
 
-  render(..._args: any[]): void {
+  render(_args: P): void {
     for (const object of this.objects) {
       object.callRender(_args)
     }